+++ lib/resque/worker.rb
@@ -371,15 +371,19 @@ module Resque
 
       worker_registry.working_on self, job
 
-      fork_for_child(job, &block)
+      run_child(job, &block)
 
     ensure
       done_working
     end
 
-    def fork_for_child(job, &block)
-      @child = ChildProcess.new(self)
-      @child.fork_and_perform(job, &block)
+    def run_child(job, &block)
+      @child = ChildProcess.create(self)
+      @child.perform(job, &block)
+    rescue => e
+      puts e.inspect
+      puts e.backtrace.join("\n")
+      raise e
     ensure
       @child = nil
     end
